edit commit

+++ src/main/java/com/expense/board/controller/ExpenseController.java
@@ -1,7 +1,9 @@
 package com.expense.board.controller;
 
+import com.expense.board.exception.ResourceNotFoundException;
 import com.expense.board.model.Expense;
 import com.expense.board.repository.ExpenseRepo;
+import com.google.common.collect.Iterables;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -9,22 +11,30 @@ import org.springframework.web.bind.annotation.*;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
-
+@CrossOrigin(origins="http://localhost:3000")
 @RestController
 @RequestMapping("/api/v1")
 public class ExpenseController
 {
     @Autowired
-    private ExpenseRepo expenseRepo;
+     ExpenseRepo expenseRepo;
     @GetMapping("/expenses")
     Iterable<Expense> getExpenses()
         {
+            if((Iterables.size(expenseRepo.findAll())<=0))
+            {
+                throw new ResourceNotFoundException("Data not available");
+            }
             return expenseRepo.findAll();
         }
     @DeleteMapping("/expenses/{id}")
-    ResponseEntity<?> deleteExpense(@PathVariable Long id){
+    public void deleteExpense(@PathVariable(value="id") Long id) throws ResourceNotFoundException
+    {
+        if (expenseRepo.findById(id).isEmpty()) {
+            throw new ResourceNotFoundException("Expense with ID " + id + " does not exist.");
+        }
         expenseRepo.deleteById(id);
-        return ResponseEntity.ok().build();
+
     }
 
     @PostMapping("/expenses")
